+++ poetry/masonry/metadata.py
@@ -52,7 +52,10 @@ class Metadata:
         meta.home_page = package.homepage or package.repository_url
         meta.author = package.author_name
         meta.author_email = package.author_email
-        meta.license = package.license.id
+
+        if package.license:
+            meta.license = package.license.id
+
         meta.classifiers = package.all_classifiers
 
         # Version 1.2